[#42] test: 리뷰 관련 로직 분리에 따른 테스트 분리
