+++ src/test/java/com/itda/apiserver/controller/ProductControllerTest.java
@@ -2,10 +2,8 @@ package com.itda.apiserver.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.itda.apiserver.domain.MainCategory;
-import com.itda.apiserver.domain.Product;
 import com.itda.apiserver.domain.Role;
 import com.itda.apiserver.domain.User;
-import com.itda.apiserver.dto.AddReviewRequestDto;
 import com.itda.apiserver.dto.AddproductRequestDto;
 import com.itda.apiserver.jwt.TokenProvider;
 import com.itda.apiserver.repository.MainCategoryRepository;
@@ -23,8 +21,6 @@ import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Optional;
 
 import static org.hamcrest.Matchers.hasSize;
@@ -163,48 +159,4 @@ class ProductControllerTest {
         return addproductRequestDto;
     }
 
-    @Test
-    @DisplayName("제품 추가 기능 테스트")
-    void addReview() throws Exception {
-
-        AddReviewRequestDto addReviewRequest = createAddReviewRequestDto();
-        String token = "Bearer " + tokenProvider.createToken(1L);
-        Product product = Product.builder()
-                .title("맛있는 사과")
-                .price(10000)
-                .build();
-
-        User user = User.builder()
-                .name("roach")
-                .email("honux")
-                .phone("01000000000")
-                .role(Role.SELLER)
-                .password("1234@@@")
-                .account("110-440-1104123")
-                .build();
-
-        when(productRepository.findById(anyLong())).thenReturn(Optional.of(product));
-        when(userRepository.findById(anyLong())).thenReturn(Optional.of(user));
-
-        mockMvc.perform(post("/api/products/1/review")
-                .header(HttpHeaders.AUTHORIZATION, token)
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(objectMapper.writeValueAsString(addReviewRequest)))
-                .andExpect(status().isOk())
-                .andDo(print());
-    }
-
-    private AddReviewRequestDto createAddReviewRequestDto() {
-
-        AddReviewRequestDto addReviewRequest = new AddReviewRequestDto();
-
-        List<String> imageList = new ArrayList<>();
-        imageList.add("image url1");
-        imageList.add("image url2");
-
-        addReviewRequest.setContents("신선하고 맛있어요");
-        addReviewRequest.setImages(imageList);
-
-        return addReviewRequest;
-    }
 }
